server/requirements.txt
fastapi
uvicorn
websockets[client]
numpy
pyyaml
torch
transformers
accelerate
huggingface_hub
requests
